Refactor : 1차 전체 리펙토링
