+++ src/main/java/com/formmaker/fff/common/response/security/UserDetailsServiceImpl.java
@@ -1,0 +1,27 @@
+package com.formmaker.fff.common.response.security;
+
+
+import com.formmaker.fff.user.entity.User;
+import com.formmaker.fff.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserDetailsServiceImpl implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String loginId) throws UsernameNotFoundException {
+
+        User user = userRepository.findByLoginId(loginId)
+                .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
+
+        return new UserDetailsImpl(user, user.getLoginId(), user.getPassword());
+    }
+
+}
