--- src/main/java/com/formmaker/fff/question/QuestionRepository.java
@@ -1,14 +1,0 @@
-package com.formmaker.fff.question;
-
-import com.formmaker.fff.survey.Survey;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface QuestionRepository extends JpaRepository<Question, Long> {
-    List<Question> findAllBySurveyId(Long surveyId);
-    void deleteAllBySurveyId(Long surveyId);
-
-    Question findAllById(Long questionId);
-}
