--- src/main/java/com/formmaker/fff/reply/ReplyRepository.java
@@ -1,6 +1,0 @@
-package com.formmaker.fff.reply;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReplyRepository extends JpaRepository<Reply, Long> {
-}
