+++ src/main/java/com/formmaker/fff/reply/repository/ReplyRepository.java
@@ -1,0 +1,7 @@
+package com.formmaker.fff.reply.repository;
+
+import com.formmaker.fff.reply.entity.Reply;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReplyRepository extends JpaRepository<Reply, Long> {
+}
