+++ src/main/java/com/formmaker/fff/stats/service/StatsService.java
@@ -1,0 +1,9 @@
+package com.formmaker.fff.stats.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class StatsService {
+}
