--- src/main/java/com/formmaker/fff/survey/request/ReplyRequest.java
@@ -1,17 +1,0 @@
-package com.formmaker.fff.survey.request;
-
-import com.formmaker.fff.common.type.QuestionTypeEnum;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-public class ReplyRequest {
-    private Long questionId;
-    private QuestionTypeEnum questionType;
-    private Integer questionNum;
-    private Integer choice;
-    private Integer selectValue;
-    private String descriptive;
-    private List<Integer> rank;
-}
