--- src/main/java/com/formmaker/fff/survey/response/AnswerResponse.java
@@ -1,13 +1,0 @@
-package com.formmaker.fff.survey.response;
-
-import com.formmaker.fff.common.type.AnswerTypeEnum;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class AnswerResponse {
-    private Integer answerNum;
-    private AnswerTypeEnum answerType;
-    private String data;
-}
