--- src/main/java/com/formmaker/fff/user/UserRepository.java
@@ -1,15 +1,0 @@
-package com.formmaker.fff.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByLoginId(String loginId);
-
-    Optional<User> findByUsername(String username);
-
-    Optional<User> findByEmail(String email);
-}
